+++ drivers/postgres/schema/internal.go
@@ -1,4 +1,4 @@
-package postgres
+package schema
 
 import "strings"
