+++ drivers/postgres/schema/internal_test.go
@@ -1,4 +1,4 @@
-package postgres
+package schema
 
 import (
 	"testing"
